refactor(sidebar): move Sidebar from packages/ui to renderer as business component
Per review feedback: packages/ui should only contain basic/primitive UI
components. The Sidebar contains business concepts (miniApp tabs, docked
tabs, cherry-glow styling) and belongs in the main project.

- Move 7 files from packages/ui/composites/Sidebar/ to
  src/renderer/src/components/Sidebar/
- Update Tooltip.tsx import: relative path → @cherrystudio/ui
- Update app/Sidebar.tsx import: @cherrystudio/ui → ../Sidebar
- Remove Sidebar exports from packages/ui/index.ts
