+++ src/renderer/src/components/Sidebar/constants.ts
