+++ src/renderer/src/components/Sidebar/index.ts
