+++ src/renderer/src/components/Sidebar/types.ts
