+++ src/renderer/src/components/Sidebar/useSidebarDrag.ts
