+++ src/renderer/src/components/Sidebar/useSidebarTabs.ts
